Update analogReference.adoc
The 1.65V and 2.23V are not generated from a build in reference. They are only calculated by the devider 1/2 and 1/1.48 from the system Voltage. People who want to measure the battery voltage in a system with less than 3.3 Volt have to use the true 1V ref. If don't write this down here correct it can become quite frustrating and timeconsuming to try to measure a bat voltage with the so called 1.65 build in reference... Maybe you will find better words to explain it to the Arduino Community but the wording "build in reference" should only be used for the 1V reference!